LayoutTests/fetch/header-constructor-is-array-expected.txt

This test should create a Headers with strange Array, but the constructor should recognize it as an Array.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS headers.get('hello') is "world"
PASS headers.get('hello') is "world"
PASS successfullyParsed is true

TEST COMPLETE
